refactor: Rename userReducers to userInfoSlice
